Fix error: Gtk-Message: Failed to load module "canberra-gtk-module"
GTK+ is a toolkit for creating GUIs, most of the default Ubuntu desktop,
and the GNOME desktop environment, is developed using it. libcanberra is
a library for playing sounds. libcanberra-gtk-module, then, is the module
that takes GUI events from GTK+ (like pressing a button, or minimizing a
window) and plays some configured sound.

The other error we get `Gtk-Message: Failed to load module "topmenu-gtk-module"`
Cannot be fixed now since topmenu is not in nixos, however there is a
request to add it. We should follow up here https://github.com/NixOS/nixpkgs/issues/73811